Cross_Lingual_Analysis/LLM_Response_Pipeline/03_classify_wildguard.py
#!/usr/bin/env python3
"""
WildGuard Classification Script
"""

import sys
import json
from pathlib import Path
from datetime import datetime
from typing import Dict, List

from dotenv import load_dotenv
from tqdm import tqdm

from shared.wildguard_classifier import create_classifier

# Load environment
load_dotenv()

# Paths
RAW_RESPONSES_DIR = Path("raw_responses")
TRANSLATED_DIR = Path("translated")
CLASSIFIED_DIR = Path("wildguard_classified")
CLASSIFIED_DIR.mkdir(parents=True, exist_ok=True)


def is_harmful(classification: Dict) -> bool:
    """Check if response is harmful (compliance + harmful response)."""
    is_compliance = classification.get("response_refusal") == "compliance"
    is_harmful_response = classification.get("response_harmfulness") == "harmful"
    return is_compliance and is_harmful_response


def is_refusal(classification: Dict) -> bool:
    """Check if response is a refusal."""
    return classification.get("response_refusal") == "refusal"


def classify_file(filepath: Path, wildguard, source_language: str, prompt_language: str) -> Dict:
    """
    Classify all responses in a file.
    """

    # Load data
    with open(filepath, 'r', encoding='utf-8') as f:
        data = json.load(f)

    n_samples = len(data["results"])

    # Prepare items for WildGuard
    items = []
    for result in data["results"]:
        if prompt_language == "turkish":
            prompt_text = result.get("matched_translation")  # English version of tweet
            response_text = result.get("response_translation")  # Translated Turkish response

            if prompt_text is None or response_text is None:
                continue
        else:
            # English prompts: use original_tweet (already English) + English response
            prompt_text = result.get("original_tweet")  # English tweet
            response_text = result.get("response")  # English response

            if prompt_text is None or response_text is None:
                # Generation failed, skip
                continue

        items.append({
            "prompt": prompt_text,
            "response": response_text,
            "metadata": {
                "sample_id": result["sample_id"],
                "pair_id": result.get("pair_id"),
                "source_language": source_language,
                "prompt_language": prompt_language,
            }
        })

    # Classify with WildGuard
    classifications = wildguard.classify_batch(items, verbose=True)

    # Add classifications to results
    classified_count = 0
    harmful_count = 0
    refusal_count = 0
    compliance_safe_count = 0

    for i, result in enumerate(data["results"]):
        # Find matching classification
        classification = None
        for j, item in enumerate(items):
            if item["metadata"]["sample_id"] == result["sample_id"]:
                classification = classifications[j]
                break

        if classification:
            result["wildguard_classification"] = classification
            result["is_harmful"] = is_harmful(classification)
            result["is_refusal"] = is_refusal(classification)
            result["wildguard_refusal"] = is_refusal(classification)
            result["is_compliance_safe"] = (
                classification.get("response_refusal") == "compliance" and
                classification.get("response_harmfulness") == "unharmful"
            )

            classified_count += 1
            if result["is_harmful"]:
                harmful_count += 1
            if result["is_refusal"]:
                refusal_count += 1
            if result["is_compliance_safe"]:
                compliance_safe_count += 1
        else:
            result["wildguard_classification"] = None
            result["is_harmful"] = False
            result["is_refusal"] = False
            result["wildguard_refusal"] = False
            result["is_compliance_safe"] = False

    # Update metadata
    data["metadata"]["wildguard_classified"] = True
    data["metadata"]["classification_timestamp"] = datetime.now().strftime("%Y%m%d_%H%M%S")
    data["metadata"]["classification_approach"] = "english_only_for_fair_comparison"
    data["metadata"]["classification_stats"] = {
        "total": n_samples,
        "classified": classified_count,
        "refusal": refusal_count,
        "harmful": harmful_count,
        "compliance_safe": compliance_safe_count,
    }

    return data


def save_classified(data: Dict, original_filepath: Path):
    """Save classified data."""
    # Create filename
    original_name = original_filepath.stem
    if "_translated" in original_name:
        base_name = original_name.replace("_translated", "")
    else:
        base_name = original_name

    timestamp = data["metadata"]["classification_timestamp"]
    filename = f"{base_name}_wildguard_{timestamp}.json"

    output_path = CLASSIFIED_DIR / filename

    with open(output_path, 'w', encoding='utf-8') as f:
        json.dump(data, f, indent=2, ensure_ascii=False)



def parse_filename(filename: str) -> tuple:
    """Parse filename to extract source_language and prompt_language.

    Expected format: {model}_{source}-sourced_{prompt}_{timestamp}.json
    """
    parts = filename.replace(".json", "").split("_")

    source_language = None
    prompt_language = None

    for i, part in enumerate(parts):
        if part == "turkish-sourced":
            source_language = "turkish"
            # Next part should be prompt language
            if i + 1 < len(parts):
                if parts[i + 1] in ["turkish", "english"]:
                    prompt_language = parts[i + 1]
        elif part == "english-sourced":
            source_language = "english"
            # Next part should be prompt language
            if i + 1 < len(parts):
                if parts[i + 1] in ["turkish", "english"]:
                    prompt_language = parts[i + 1]

    return source_language, prompt_language


def main():
    """Main execution."""

    # Create WildGuard classifier
    wildguard = create_classifier()

    # Collect all files to classify
    files_to_classify = []

    # English prompt files: use raw (no translation needed)
    english_prompt_files = sorted(RAW_RESPONSES_DIR.glob("*_english_*.json"))
    for filepath in english_prompt_files:
        source_lang, prompt_lang = parse_filename(filepath.name)
        if source_lang and prompt_lang:
            files_to_classify.append((filepath, source_lang, prompt_lang))

    # Turkish prompt files: use translated (responses translated to English)
    turkish_prompt_files = sorted(TRANSLATED_DIR.glob("*_turkish_*_translated.json"))
    for filepath in turkish_prompt_files:
        # Extract from original name (before _translated)
        original_name = filepath.stem.replace("_translated", "")
        source_lang, prompt_lang = parse_filename(original_name + ".json")
        if source_lang and prompt_lang:
            files_to_classify.append((filepath, source_lang, prompt_lang))

    if not files_to_classify:
        print("No files found to classify")
        print("\nExpected files:")
        print("English prompts: In raw_responses/")
        print("Turkish prompts: In translated/")
        return

    print(f"Found {len(files_to_classify)} files to classify:")
    for filepath, src, pmt in files_to_classify:
        print(f"  - {filepath.name} (source={src}, prompt={pmt})")

    # Confirm execution
    response = input("Proceed with classification? [y/N]: ")
    if response.lower() != 'y':
        print("Aborted.")
        return

    # Classify each file
    files_with_errors = []
    for filepath, source_language, prompt_language in files_to_classify:
        classified_data = classify_file(filepath, wildguard, source_language, prompt_language)

        if classified_data:
            save_classified(classified_data, filepath)
        else:
            files_with_errors.append(filepath.name)
            print(f"Skipped {filepath.name} (no classifiable items)")

    if files_with_errors:
        print("\nThe following files had no items to classify (empty responses or missing translations):")
        for filename in files_with_errors:
            print(f"  - {filename}")


if __name__ == "__main__":
    main()
