pyproject.toml
[project]
name = "gemmaguard"
version = "0.1.0"
description = "This repo houses all the code and the required packages for analyzing cross lingual safety gaps"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "anthropic>=0.71.0",
    "bitsandbytes>=0.48.2",
    "datasets>=4.2.0",
    "fastparquet>=2024.11.0",
    "google-cloud-storage>=3.6.0",
    "huggingface-hub>=0.35.3",
    "matplotlib>=3.10.7",
    "numpy>=2.3.3",
    "openai>=2.5.0",
    "optuna>=4.6.0",
    "pandas>=2.3.3",
    "peft>=0.18.0",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.5",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.0",
]
